requirements.txt
captcha==0.6.0
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
